+++ src/generators/generator_circle_shape3d.h
@@ -73,7 +73,7 @@ public:
 	void set_projection_collision_mask(int mask);
 	int get_projection_collision_mask() const { return projection_collision_mask; }
 
-	void perform_generation(uint64_t initial_time_usec, int time_budget_ms) override;
+	void perform_generation(uint64_t initial_time_usec, double time_budget_ms) override;
 
 protected:
 	static void _bind_methods();
